Fixed compiler warnings in brc

+++ _studio/shared/umc/codec/brc/src/umc_video_brc.cpp
@@ -4,7 +4,7 @@
 //     This software is supplied under the terms of a license agreement or
 //     nondisclosure agreement with Intel Corporation and may not be copied
 //     or disclosed except in accordance with the terms of that agreement.
-//          Copyright(c) 2010-2012 Intel Corporation. All Rights Reserved.
+//          Copyright(c) 2010-2016 Intel Corporation. All Rights Reserved.
 //
 //
 //                     bitrate control interface
@@ -94,7 +94,7 @@ Status VideoBrc::CheckCorrectParams_MPEG2(VideoBrcParams *inBrcParams, VideoBrcP
   if (outBrcParams)
     *outBrcParams = *inBrcParams;
 
-  if (bufferSizeBytes > IPP_MAX_32U >> 3)
+  if (static_cast<unsigned int>(bufferSizeBytes) > IPP_MAX_32U >> 3)
     bufferSizeBytes = IPP_MAX_32U >> 3;
 
   bufSizeBits = bufferSizeBytes << 3;
